EF-353,EF-382,EF-394,EF-395,EF-396,EF-397,EF-402,EF-410,EF-413,EF-422,EF-423,EF-424,EF-427,EF-428: close small native-LINQ translation gaps
Closes 14 "small feature gap" tickets from the EF-322 native LINQ rewrite
epic's classification survey (EF-247, EF-335, EF-362, EF-405 were found
already shipped or externally blocked and are not part of this change).
Each closes a narrow gap in the native MQL query translator so more LINQ
shapes execute as native aggregation pipelines instead of falling back to
(or crashing on) the driver's own LINQ provider.

- EF-394: composite primary-key member access resolves natively via the
  dotted `_id.<name>` path instead of declining.
- EF-410: a widening numeric `Convert` is now admitted by the tier-2
  bare-projection gate.
- EF-402: a `Nullable<T>.Value` projection read now resolves through its
  own property's value converter instead of silently reading the raw
  stored value.
- EF-413: `MongoAggregationExpressionRenderer` gained renderer arms for
  `MongoInExpression`/`MongoUnaryExpression`, closing a value-converted-bool
  truthiness hazard this uncovered in the process.
- EF-396: `Not` over an unsupported subtree (e.g. a field-to-field
  comparison, or a conjunction over bare bool fields) now renders via
  `$expr` instead of declining, with a matching truthiness guard for bare
  bool operands of `&&`/`||`.
- EF-382: `arrayField.Contains(constant)` now translates natively via a
  dedicated `MongoArrayContainsExpression` node (not reused as `$eq`, to
  avoid a whole-array-equality mis-render if used as a boolean value).
- EF-423: non-TPH `OfType` disposition pinned by a regression test (found
  to already behave correctly in every mode; no production fix needed).
- EF-424: dotted-path resolution now composes correctly through an
  intermediate owned-collection hop from a non-root (element-scoped) scope.
- EF-427: stranded bare-collection `First`/`FirstOrDefault`/`Single`/
  `SingleOrDefault`/`Any` projection leaves rebuilt against `Enumerable`
  equivalents instead of throwing; filtered `Count(pred)` reachability
  widened to a pure arithmetic/cast spine.
- EF-353,EF-428: owned `SelectMany` whole-element sentinel collision fixed
  (owner-key/ordinal sentinels nested under one reserved wrapper field so
  an ordinary stored property can never collide with them), and a nested
  owned member on the re-rooted element now binds against the element's
  own projection instead of the outer query's.
- EF-422: `SelectMany` cross-scope computed projection leaves and the
  inner-Select-form computed-leaf binder.
- EF-395: bare-projection operands of set operations (Union/Intersect/
  Except) and of `Distinct` now go native.
- EF-397: a reducer (First/Single/etc.) composed after `Take`/`Skip` now
  composes its own `$limit` instead of declining (consecutive `$limit`
  stages narrow monotonically); a collection `Include` combined with a set
  operation now goes native via a `MongoSelectLowerer` pipeline-stage
  reorder that defers `$lookup` until after the set-op stage (the literal
  gate-relaxation fix originally proposed would have shipped silent wrong
  data — proven by mutation testing during review).

Two genuine silent-wrong-data regressions were caught and fixed during
review before landing: a value-converted bool's MQL truthiness diverging
from its CLR value when negated (both directly under `Not` and as a bare
operand of `&&`/`||`), surfaced across the EF-413/EF-396 work.

Executed via subagent-driven development with a fresh implementer and an
independent reviewer per task, plus a final whole-branch review; every
task from EF-413 onward required at least one review-driven fix round
except the final task (EF-397), which passed clean on first review.

+++ tests/MongoDB.EntityFrameworkCore.UnitTests/Query/NativeTranslation/SlotPopulationTests.cs
@@ -212,12 +212,16 @@ public class SlotPopulationTests
     // FLIPPED by EF-322 step 3a. This test used to use a BARE scalar body (`c => c.Name`) as its example of a
     // non-representable projection; that shape is now native (see Bare_scalar_projection_is_native above), so the
     // example has to be one the binder still declines or the test would be asserting the opposite of the truth.
-    // A type-CHANGING cast leaf is the closest still-declining shape — the computed long tail, unrelated to the
-    // bare/wrapped boundary — so what this test pins is unchanged: a declined projection drives Route to Fallback.
+    // A WIDENING cast (`(long)c.Age`) was this test's example through EF-410; that shape is now ALSO native (a
+    // widening Convert is admitted as a bare MongoFieldExpression — see NativeProjectionBinder's tier-2 gate),
+    // so the example moved again, to a NARROWING cast with no admissible MQL conversion operator ($toShort does
+    // not exist — see MongoConvertExpression.ToOperatorFor). That is the still-declining computed long tail,
+    // unrelated to the bare/wrapped boundary — so what this test pins is unchanged: a declined projection drives
+    // Route to Fallback.
     [Fact]
     public void A_declined_projecting_Select_is_not_native_representable()
     {
-        var mongoQ = TranslateToMongoQuery<Customer>(q => q.Select(c => (long)c.Age));
+        var mongoQ = TranslateToMongoQuery<Customer>(q => q.Select(c => (short)c.Age));
 
         Assert.Equal(NativeRoute.Fallback, mongoQ.Select.Route);
         Assert.Empty(mongoQ.Select.Projection);
@@ -336,9 +340,12 @@ public class SlotPopulationTests
     }
 
     [Fact]
+    // A WIDENING cast member (`(long)c.Age`) was this test's example through EF-410; that shape is now native
+    // (see NativeCastTests.Widening_cast_projection_leaf_now_goes_native), so this uses a NARROWING cast to a
+    // target with no admissible MQL conversion operator ($toShort does not exist) instead, which still declines.
     public void Cast_member_projection_is_not_native()
     {
-        var mongoQuery = TranslateToMongoQuery<Customer>(q => q.Select(c => new { Position = (long)c.Age }));
+        var mongoQuery = TranslateToMongoQuery<Customer>(q => q.Select(c => new { Position = (short)c.Age }));
 
         Assert.Equal(NativeRoute.Fallback, mongoQuery.Select.Route);
         Assert.Empty(mongoQuery.Select.Projection);
